requirements.txt
streamlit
langchain-core
python-dotenv
langchain-google-genai
